requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
darkskylib==0.3.91
decorator==5.1.1
Django==5.0.6
exceptiongroup==1.2.1
executing==2.0.1
future==1.0.0
gunicorn==22.0.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
json2html==1.3.0
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
numpy==2.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
